refactor server code into mvc and add controllers for review
