+++ package.json
@@ -11,8 +11,8 @@
     "server-cassandra": "nodemon server/server-cassandra.js",
     "server-postgres": "nodemon server/server-postgres.js",
     "seed-cassandra": "node server/db/cassandra/seed.js",
-    "seed-postgres": "node server/db/postgres/seed.js",
-    "seed-review": "node server/db/review/seed.js",
+    "seed-postgres": "node server/db/postgres/seed.js && npm run gzip postgres",
+    "seed-review": "node server/db/review/seed.js && npm run gzip review",
     "gzip-postgres": "gzip server/db/postgres/restaurant_category.csv && gzip server/db/postgres/category.csv && gzip server/db/postgres/restaurant.csv",
     "gzip-review": "gzip server/db/review/review.csv && gzip server/db/review/review_dish.csv",
     "start": "node server/db/mongo/seed.js && npm run server",
