--- server/db/mongo/index.js
@@ -1,5 +1,0 @@
-const mongoose = require('mongoose');
-
-const db = mongoose.connection;
-
-module.exports = db;
