+++ server/db/review/query.sql
@@ -1,0 +1,4 @@
+\connect header root;
+
+CREATE INDEX review_index ON review_dish(review_id);
+CREATE INDEX dish_index ON review_dish(dish_id);
