+++ server/db/review/schema.sql
@@ -1,7 +1,6 @@
 -- STEPS:
 -- 1. npm run seed-review
--- 2. npm run gzip-review
--- 3. psql postgres -f ./server/db/review/schema.sql
+-- 2. psql postgres -f ./server/db/review/schema.sql
 
 -- CREATE ROLE root WITH SUPERUSER;
 \connect header root;
