+++ server/server-postgres.js
@@ -1,23 +1,23 @@
 const express = require('express');
-const db = require('./controllers');
+const controllers = require('./controllers/postgres');
 
 const app = express();
 const port = 3003;
 
 app.use(express.json());
 
-app.get('/api/:id/header', db.getHeader);
-app.get('/api/:id/review', db.getReview);
-app.get('/api/:id/category', db.getCategory);
+app.get('/api/:id/header', controllers.getHeader);
+// app.get('/api/:id/review', controllers.getReview);
+// app.get('/api/:id/category', controllers.getCategory);
 
 
-app.post('/api/:id/header', db.postHeader);
-app.post('/api/:id/category', db.postCategory);
+// app.post('/api/:id/header', controllers.postHeader);
+// app.post('/api/:id/category', controllers.postCategory);
 
-app.patch('/api/:id/header', db.patchHeader);
-app.patch('/api/:id/category', db.patchCategory);
+// app.patch('/api/:id/header', controllers.patchHeader);
+// app.patch('/api/:id/category', controllers.patchCategory);
 
-app.delete('/api/:id/header', db.deleteHeader);
-app.delete('/api/:id/category', db.deleteCategory);
+// app.delete('/api/:id/header', controllers.deleteHeader);
+// app.delete('/api/:id/category', controllers.deleteCategory);
 
 app.listen(port, () => console.log(`Server listening on port ${port}`));
